Add Mesh class and vertex+texture structs
Mesh class is a class that takes it's vertices, indices and textures and
is able to self-render given a specific shader. It internally puts
everything in VAOs/VBOs/EBOs.

A vertex is defined by its position, its normal and its texture
coordinates. A texture is defined by its ID, its type, and the path (as
an aiString).